Extracted room drawing code to its own file.
